AMBARI-23438 : Fix failing metrics unit tests on trunk. (RawMetricsSourceTest)
