+++ ambari-metrics/ambari-metrics-common/pom.xml
@@ -110,10 +110,6 @@
                   <shadedPattern>org.apache.ambari.metrics.sink.relocated.jboss</shadedPattern>
                 </relocation>
                 <relocation>
-                  <pattern>net.sf.ehcache</pattern>
-                  <shadedPattern>org.apache.ambari.metrics.sink.relocated.ehcache</shadedPattern>
-                </relocation>
-                <relocation>
                   <pattern>org.apache.http</pattern>
                   <shadedPattern>org.apache.hadoop.metrics2.sink.relocated.apache.http</shadedPattern>
                 </relocation>
@@ -136,11 +132,6 @@
   </build>
 
   <dependencies>
-    <dependency>
-      <groupId>net.sf.ehcache</groupId>
-      <artifactId>ehcache</artifactId>
-      <version>2.10.0</version>
-    </dependency>
     <dependency>
       <groupId>commons-logging</groupId>
       <artifactId>commons-logging</artifactId>
